chore(cleanup): extract analytics service out of renderer
