+++ app/renderer/components/sensor/edit/Form.jsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Component } from 'react'
-import { methods, MANUAL } from '../../../services/analytics/MaxHeartRateCalculators'
-import { getMaxHeartRate } from '../../../services/analytics/'
+import { methods, MANUAL } from '../../../../services/analytics/MaxHeartRateCalculators'
+import { getMaxHeartRate } from '../../../../services/analytics/'
 import { MenuItem, Grid, TextField, InputAdornment } from '@material-ui/core'
 import Coefficients from './Coefficients'
 import { withStyles } from '@material-ui/core/styles'
