+++ app/services/analytics/CalorieBurnPerHourCalculators.js
