+++ app/services/analytics/MaxHeartRateCalculators.js
