+++ app/services/analytics/calorie_burn_per_hour_calculators.test.js
