+++ app/services/analytics/max_heart_rate_calculators_test.test.js
